intermediate fixes
I did not run unit tests for this commit, and for the last couple
versions (last commits for sure) I didn't pay any attention to
breakages. There are a lot of parts that I no longer need, because I
provided better and more streamlined implementation in
nimskull (specifically colored strings), and I don't want to expend
energy maintaining this library.

- fix: =clformat= for binary renders zero as a single ~0~ instead of
  completely empty string
- fix: show string description at mismatch position
- feature: split closure int proc + env, or just proc
- feature: string templates for formatting dates using ISO
  format (~isoDateFmtMsec~)
- feature: instantiation depth for current source directory
- fix: provide default dummy implementation for procedure writing and
  reading, for json serialization. Procedure is always serialized as
  null value - user can override this later, if they choose to do so.
- src/hmisc/other/blockfmt.nim:26 :: use ~high()~ instead of hardcoded
  value to avoid compilation issues when targeting 32-bit systems